refactor: aislado documentviewer en su propio organismo + manejo de estados ahora se hace con un reducer
