+++ components/organisms/document-viewver/hooks/useDocumentViewerContext.ts
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import { DocumentContext } from "../context/context";
+
+export function useDocumentViewerContext() {
+  return useContext(DocumentContext);
+}
