+++ contexts/language-context.tsx
@@ -1,45 +1,22 @@
-"use client"
+"use client";
 
-import * as React from "react"
-import { translations, type Language } from "@/lib/translations"
+import * as React from "react";
+import { type Language } from "@/lib/translations";
 
 type LanguageContextType = {
-  language: Language
-  setLanguage: (lang: Language) => void
-  t: (key: string) => string
-}
+  language: Language;
+  setLanguage: (lang: Language) => void;
+  t: (key: string) => string;
+};
 
-const LanguageContext = React.createContext<LanguageContextType | undefined>(undefined)
-
-export function LanguageProvider({ children }: { children: React.ReactNode }) {
-  const [language, setLanguage] = React.useState<Language>("es")
-
-  const t = React.useCallback(
-    (key: string) => {
-      const keys = key.split(".")
-      let value: any = translations[language]
-
-      for (const k of keys) {
-        if (value && typeof value === "object") {
-          value = value[k]
-        } else {
-          return key
-        }
-      }
-
-      return value || key
-    },
-    [language],
-  )
-
-  return <LanguageContext.Provider value={{ language, setLanguage, t }}>{children}</LanguageContext.Provider>
-}
+export const LanguageContext = React.createContext<
+  LanguageContextType | undefined
+>(undefined);
 
 export function useLanguage() {
-  const context = React.useContext(LanguageContext)
+  const context = React.useContext(LanguageContext);
   if (context === undefined) {
-    throw new Error("useLanguage must be used within a LanguageProvider")
+    throw new Error("useLanguage must be used within a LanguageProvider");
   }
-  return context
+  return context;
 }
-
