+++ layouts/default/container.tsx
@@ -1,0 +1,5 @@
+export function Container({ children }: { children: React.ReactNode }) {
+  return (
+    <div className="flex min-h-screen flex-col bg-background">{children}</div>
+  );
+}
