+++ layouts/default/header/index.tsx
@@ -1,0 +1,14 @@
+import { HorizontalMenu } from "./horizontal-menu";
+import { Logo } from "./logo";
+
+export function Header() {
+  return (
+    <header className="border-b">
+      <div className="container flex h-14 items-center justify-between px-4">
+        <Logo />
+
+        <HorizontalMenu />
+      </div>
+    </header>
+  );
+}
