+++ package.json
@@ -45,6 +45,7 @@
     "embla-carousel-react": "8.5.1",
     "input-otp": "1.4.1",
     "js-yaml": "latest",
+    "lodash": "^4.17.21",
     "lucide-react": "^0.454.0",
     "next": "15.2.4",
     "next-themes": "^0.4.4",
@@ -64,6 +65,7 @@
   },
   "devDependencies": {
     "@types/js-yaml": "^4.0.9",
+    "@types/lodash": "^4.17.16",
     "@types/node": "^22",
     "@types/prismjs": "^1.26.5",
     "@types/react": "^19",
