+++ providers/language-provider.tsx
@@ -1,0 +1,33 @@
+"use client";
+
+import { type Language, translations } from "@/lib/translations";
+import * as React from "react";
+import { LanguageContext } from "../contexts/language-context";
+
+export function LanguageProvider({ children }: { children: React.ReactNode }) {
+  const [language, setLanguage] = React.useState<Language>("es");
+
+  const t = React.useCallback(
+    (key: string) => {
+      const keys = key.split(".");
+      let value: any = translations[language];
+
+      for (const k of keys) {
+        if (value && typeof value === "object") {
+          value = value[k];
+        } else {
+          return key;
+        }
+      }
+
+      return value || key;
+    },
+    [language]
+  );
+
+  return (
+    <LanguageContext.Provider value={{ language, setLanguage, t }}>
+      {children}
+    </LanguageContext.Provider>
+  );
+}
